refactor encoder
