+++ sktime_dl/classification/_encoder.py
@@ -1,7 +1,7 @@
 __author__ = "James Large"
 
 from sktime_dl.deeplearning.base.estimators import BaseDeepClassifier
-from sktime_dl.deeplearning.encoder._base import EncoderNetwork
+from sktime_dl.networks._encoder import EncoderNetwork
 from sktime_dl.utils import check_and_clean_data, \
     check_and_clean_validation_data
 from tensorflow import keras
