+++ sktime_dl/deeplearning/tests/test_accuracy.py
@@ -12,7 +12,7 @@ from flaky import flaky
 from sktime.datasets import load_italy_power_demand
 
 from sktime_dl.classification import CNNClassifier
-from sktime_dl.deeplearning import EncoderClassifier
+from sktime_dl.classification import EncoderClassifier
 from sktime_dl.classification import FCNClassifier
 from sktime_dl.classification import InceptionTimeClassifier
 from sktime_dl.classification import MCDCNNClassifier
