+++ sktime_dl/deeplearning/tests/test_classifiers.py
@@ -1,8 +1,8 @@
 from sktime.datasets import load_basic_motions
 from sktime.datasets import load_italy_power_demand
 
-from sktime_dl.deeplearning import LSTMFCNClassifier
-from sktime_dl.deeplearning.cntc._classifier import CNTCClassifier
+from sktime_dl.classification import LSTMFCNClassifier
+from sktime_dl.classification import CNTCClassifier
 from sktime_dl.utils.model_lists import SMALL_NB_EPOCHS
 from sktime_dl.utils.model_lists import construct_all_classifiers
