+++ sktime_dl/networks/_encoder.py
