+++ sktime_dl/regression/_encoder.py
@@ -3,7 +3,7 @@ __author__ = "James Large, Withington"
 from tensorflow import keras
 
 from sktime_dl.deeplearning.base.estimators import BaseDeepRegressor
-from sktime_dl.deeplearning.encoder._base import EncoderNetwork
+from sktime_dl.networks._encoder import EncoderNetwork
 from sktime_dl.utils import check_and_clean_data, \
     check_and_clean_validation_data
